+++ styx/src/engine/helper/geom/CPlane.cpp
@@ -21,24 +21,24 @@ const glm::vec3 &CPlane::Normal() const
 	return( m_normal );
 }
 
-void CPlane::SetDistance( float distance )
+void CPlane::SetDistance( f16 distance )
 {
 	m_distance = distance;
 }
 
-float CPlane::Distance() const
+f16 CPlane::Distance() const
 {
 	return( m_distance );
 }
 
 void CPlane::Normalize()
 {
-	const float length = 1.0f / glm::length( m_normal );
+	const f16 length = 1.0f / glm::length( m_normal );
 	m_normal	*= length;
 	m_distance	*= length;
 }
 
-float CPlane::DistanceToPlane( const glm::vec3 &point ) const
+f16 CPlane::DistanceToPlane( const glm::vec3 &point ) const
 {
 	return( glm::dot( m_normal, point ) + m_distance );
 }
