+++ styx/src/engine/helper/geom/CPlane.hpp
@@ -2,6 +2,8 @@
 
 #include <glm/glm.hpp>
 
+#include "src/engine/helper/Types.hpp"
+
 class CPlane
 {
 public:
@@ -12,14 +14,14 @@ public:
 	void SetNormal( const glm::vec3 &normal );
 	const glm::vec3 &Normal() const;
 
-	void SetDistance( float distance );
-	float Distance() const;
+	void SetDistance( f16 distance );
+	f16 Distance() const;
 
 	void Normalize();
 
-	float DistanceToPlane( const glm::vec3 &point ) const;
+	f16 DistanceToPlane( const glm::vec3 &point ) const;
 
 private:
 	glm::vec3	m_normal	{ 0.0f, 0.0f, 0.0f };
-	float		m_distance	{ 0.0f };
+	f16			m_distance	{ 0.0f };
 };
